@@ -11,15 +11,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~10.0.0",
-    "@angular/cli": "^10.0.0",
-    "@angular/common": "~10.0.0",
-    "@angular/compiler": "~10.0.0",
-    "@angular/core": "~10.0.0",
-    "@angular/forms": "~10.0.0",
-    "@angular/platform-browser": "~10.0.0",
-    "@angular/platform-browser-dynamic": "~10.0.0",
-    "@angular/router": "~10.0.0",
+    "@angular/animations": "~11.2.5",
+    "@angular/cli": "^11.2.4",
+    "@angular/common": "~11.2.5",
+    "@angular/compiler": "~11.2.5",
+    "@angular/core": "~11.2.5",
+    "@angular/forms": "~11.2.5",
+    "@angular/platform-browser": "~11.2.5",
+    "@angular/platform-browser-dynamic": "~11.2.5",
+    "@angular/router": "~11.2.5",
     "bootstrap": "^4.5.0",
     "ngx-bootstrap": "^6.2.0",
     "rxjs": "~6.5.5",
@@ -27,22 +27,23 @@
     "zone.js": "~0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1000.0",
-    "@angular/compiler-cli": "~10.0.0",
-    "@types/jasmine": "~3.5.0",
+    "@angular-devkit/build-angular": "~0.1102.4",
+    "@angular/compiler-cli": "~11.2.5",
+    "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.11.1",
-    "codelyzer": "^6.0.0-next.1",
-    "jasmine-core": "~3.5.0",
+    "codelyzer": "^6.0.0",
+    "jasmine-core": "~3.6.0",
     "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~5.0.0",
+    "karma": "~6.2.0",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "~3.0.2",
-    "karma-jasmine": "~3.3.0",
+    "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
     "protractor": "~7.0.0",
+    "ts-mockito": "^2.6.1",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
-    "typescript": "~3.9.5"
+    "typescript": "~4.1.5"
   }
 }
