@@ -1,38 +1,21 @@
-  import { BrowserModule } from '@angular/platform-browser';
-import { NgModule } from '@angular/core';
-import { FormsModule } from '@angular/forms';
-import { AlertModule } from 'ngx-bootstrap/alert';
-import { HttpClientModule } from '@angular/common/http';
-import { RouterModule, Routes } from '@angular/router';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import {BrowserModule} from '@angular/platform-browser';
+import {NgModule} from '@angular/core';
+import {FormsModule} from '@angular/forms';
+import {AlertModule} from 'ngx-bootstrap/alert';
+import {HttpClientModule} from '@angular/common/http';
+import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
 
-import { AppRoutingModule } from './app-routing.module';
-import { AppComponent } from './app.component';
-import { TopicComponent } from './topic/topic.component';
-import { TopicService } from './topic/topic.service';
-import { TopicListComponent } from './topic/topic-list/topic-list.component';
-import { TopicAddComponent } from './topic/topic-add/topic-add.component';
-import { TopicDetailsComponent } from './topic/topic-details/topic-details.component';
-import { TopicPublishComponent } from './topic/topic-publish/topic-publish.component';
-
-const routes: Routes = [
-  {
-    path: 'topics',
-    component: TopicComponent
-  },
-  {
-    path: 'topics/new',
-    component: TopicAddComponent
-  },
-  {
-    path: 'topics/:name',
-    component: TopicDetailsComponent
-  },
-  {
-    path: 'topics/:name/publish',
-    component: TopicPublishComponent
-  }
-]
+import {AppRoutingModule} from './app-routing.module';
+import {AppComponent} from './app.component';
+import {TopicComponent} from './topic/topic.component';
+import {TopicService} from './topic/topic.service';
+import {TopicListComponent} from './topic/topic-list/topic-list.component';
+import {TopicAddComponent} from './topic/topic-add/topic-add.component';
+import {TopicPublishComponent} from './topic/topic-publish/topic-publish.component';
+import {PaginationComponent} from './pagination/pagination.component';
+import { HomeComponent } from './home/home.component';
+import { BreadcrumbComponent } from './breadcrumb/breadcrumb.component';
+import {TopicRecordsComponent} from './topic/topic-records/topic-records.component';
 
 @NgModule({
   declarations: [
@@ -40,8 +23,11 @@ const routes: Routes = [
     TopicListComponent,
     TopicComponent,
     TopicAddComponent,
-    TopicDetailsComponent,
+    TopicRecordsComponent,
     TopicPublishComponent,
+    PaginationComponent,
+    HomeComponent,
+    BreadcrumbComponent,
   ],
   imports: [
     BrowserModule,
@@ -49,8 +35,7 @@ const routes: Routes = [
     FormsModule,
     BrowserAnimationsModule,
     AlertModule,
-    HttpClientModule,
-    RouterModule.forRoot(routes)
+    HttpClientModule
   ],
   providers: [TopicService],
   bootstrap: [AppComponent]
