@@ -1,16 +1,27 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import {ComponentFixture, TestBed, waitForAsync} from '@angular/core/testing';
 
-import { TopicAddComponent } from './topic-add.component';
+import {TopicAddComponent} from './topic-add.component';
+import {TopicService} from '../topic.service';
+import {Router} from '@angular/router';
+import {NgForm} from '@angular/forms';
 
 describe('TopicAddComponent', () => {
   let component: TopicAddComponent;
   let fixture: ComponentFixture<TopicAddComponent>;
+  let topicServiceStub: any;
+  let routerStub: any;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
+    topicServiceStub = {};
+    routerStub = {};
     TestBed.configureTestingModule({
-      declarations: [ TopicAddComponent ]
+      declarations: [TopicAddComponent, NgForm],
+      providers: [
+        {provide: TopicService, useValue: topicServiceStub},
+        {provide: Router, useValue: routerStub},
+      ]
     })
-    .compileComponents();
+      .compileComponents();
   }));
 
   beforeEach(() => {
