@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { TopicListComponent } from './topic-list.component';
 
@@ -6,7 +6,7 @@ describe('TopicListComponent', () => {
   let component: TopicListComponent;
   let fixture: ComponentFixture<TopicListComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ TopicListComponent ]
     })
